[Resources] Separate HostResources from resources.h/cc
