+++ src/framework/scene/resources.h
@@ -1,13 +1,8 @@
 #ifndef VKFRAMEWORK_SCENE_RESOURCES_H
 #define VKFRAMEWORK_SCENE_RESOURCES_H
 
-#include "framework/common.h"
+#include "framework/scene/host_resources.h"
 
-#include "framework/scene/animation.h"
-#include "framework/scene/texture.h"
-#include "framework/scene/image_data.h"
-#include "framework/scene/material.h"
-#include "framework/scene/mesh.h"
 #include "framework/renderer/fx/material/material_fx_registry.h"
 
 class Context;
@@ -19,45 +14,6 @@ class Camera;
 /* -------------------------------------------------------------------------- */
 
 namespace scene {
-
-template<typename T>
-using ResourceBuffer = std::vector<std::unique_ptr<T>>;
-
-template<typename T>
-using ResourceMap = std::unordered_map<std::string, std::unique_ptr<T>>;
-
-// ----------------------------------------------------------------------------
-// ----------------------------------------------------------------------------
-
-struct HostResources {
-  HostResources() = default;
-
-  virtual ~HostResources() = default;
-
-  virtual void setup();
-
-  virtual bool load_file(std::string_view filename);
-
-  /* --- Host Data --- */
-
-  std::vector<scene::Sampler> samplers{};
-  std::vector<ImageData> host_images{};
-  std::vector<Texture> textures{}; //
-
-  ResourceBuffer<MaterialRef> material_refs{}; //
-
-  ResourceBuffer<Mesh> meshes{}; //
-  std::vector<mat4f> transforms{};
-
-  ResourceBuffer<Skeleton> skeletons{};
-  ResourceMap<AnimationClip> animations_map{};
-
- protected:
-  DefaultTextureBinding default_bindings_{};
-};
-
-// ----------------------------------------------------------------------------
-// ----------------------------------------------------------------------------
 
 ///
 /// Holds Host and Device resources for rendering a (glTF) scene.
@@ -102,33 +58,15 @@ struct Resources : HostResources {
   }
 
  private:
-  void reset_internal_descriptors();
-
   void upload_images(Context const& context);
-
   void upload_buffers(Context const& context);
 
  public:
-  // /* --- Host Data --- */
-
-  // std::vector<ImageData> host_images{};
-  // std::vector<Texture> textures{};
-  // ResourceBuffer<MaterialRef> material_refs{};
-  // ResourceBuffer<Skeleton> skeletons{};
-  // ResourceBuffer<Mesh> meshes{};
-  // std::vector<mat4f> transforms{};
-
-  // ResourceMap<AnimationClip> animations_map{};
-
   /* --- Device Data --- */
 
   std::vector<backend::Image> device_images{};
   backend::Buffer vertex_buffer{};
   backend::Buffer index_buffer{};
-
-  uint32_t vertex_buffer_size{0u};
-  uint32_t index_buffer_size{0u};
-  uint32_t total_image_size{0u};
 
  private:
   Renderer const* renderer_ptr_{};
@@ -140,17 +78,19 @@ struct Resources : HostResources {
   backend::Buffer frame_ubo_{};
   backend::Buffer transforms_ssbo_{};
 
-  ///
-  /// Limitations : if a MaterialFx have others states changes than its alphamode
-  ///                 some might be discarded...
+  /// --------------------------
+  /// (warning)
+  /// If a MaterialFx have others states changes than its alphamode
+  /// some might be discarded...
   ///
   using SubMeshBuffer = std::vector<Mesh::SubMesh const*>;
   using FxToSubmeshesMap = std::map< MaterialFx*, SubMeshBuffer >;
+  /// --------------------------
 
-  EnumArray<FxToSubmeshesMap, MaterialStates::AlphaMode> lookups_;
+  EnumArray<FxToSubmeshesMap, MaterialStates::AlphaMode> lookups_{};
 };
 
-}  // namespace scene
+} // namespace scene
 
 /* -------------------------------------------------------------------------- */
